Numberの設定範囲を変更

+++ WpfApp1/WpfApp1/MainWindow/MainWindowViewModel.cs
@@ -57,7 +57,7 @@ namespace WpfApp1.MainWindow
         /// <summary>
         /// Number設定
         /// </summary>
-        public ReactivePropertySlim<int> Number { get; }
+        public ReactivePropertySlim<double> Number { get; }
 
         /// <summary>
         /// Number設定が変更されたか
